Move `CardinalityEstimatorTrait` to be only used by benchmarks.
This trait was needed for benchmarks in the first place, so moving it there to simplify package usability.
