+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "cardinality-estimator"
-version = "1.0.0"
+version = "1.0.1"
 edition = "2021"
 authors = ["Alex Bocharov <bocharov.alexandr@gmail.com>"]
 description = "A crate for estimating the cardinality of distinct elements in a stream or dataset."
