+++ benches/cardinality_estimator.rs
@@ -1,9 +1,9 @@
 #[global_allocator]
 static ALLOC: dhat::Alloc = dhat::Alloc;
 
-use std::hash::BuildHasherDefault;
+use std::hash::{BuildHasherDefault, Hash};
 
-use cardinality_estimator::{CardinalityEstimator, CardinalityEstimatorTrait};
+use cardinality_estimator::CardinalityEstimator;
 use criterion::measurement::WallTime;
 use criterion::{
     black_box, criterion_group, criterion_main, BenchmarkGroup, BenchmarkId, Criterion, Throughput,
@@ -43,7 +43,7 @@ fn benchmark(c: &mut Criterion) {
     let mut group = c.benchmark_group("insert");
     for &cardinality in &cardinalities {
         group.throughput(Throughput::Elements(cardinality.max(1) as u64));
-        bench_insert::<CardinalityEstimator<usize>>(&mut group, cardinality);
+        bench_insert::<CardinalityEstimatorMut>(&mut group, cardinality);
         bench_insert::<AmadeusStreamingEstimator>(&mut group, cardinality);
         bench_insert::<ProbabilisticCollections>(&mut group, cardinality);
         bench_insert::<HyperLogLog>(&mut group, cardinality);
@@ -54,7 +54,7 @@ fn benchmark(c: &mut Criterion) {
     let mut group = c.benchmark_group("estimate");
     group.throughput(Throughput::Elements(1));
     for &cardinality in &cardinalities {
-        bench_estimate::<CardinalityEstimator<usize>>(&mut group, cardinality);
+        bench_estimate::<CardinalityEstimatorMut>(&mut group, cardinality);
         bench_estimate::<AmadeusStreamingEstimator>(&mut group, cardinality);
         bench_estimate::<ProbabilisticCollections>(&mut group, cardinality);
         bench_estimate::<HyperLogLog>(&mut group, cardinality);
@@ -66,7 +66,7 @@ fn benchmark(c: &mut Criterion) {
         .iter()
         .map(|&cardinality| StatRecord {
             cardinality,
-            cardinality_estimator: measure_allocations::<CardinalityEstimator<usize>>(cardinality),
+            cardinality_estimator: measure_allocations::<CardinalityEstimatorMut>(cardinality),
             amadeus_streaming: measure_allocations::<AmadeusStreamingEstimator>(cardinality),
             probabilistic_collections: measure_allocations::<ProbabilisticCollections>(cardinality),
             hyperloglog: measure_allocations::<HyperLogLog>(cardinality),
@@ -85,7 +85,7 @@ fn benchmark(c: &mut Criterion) {
         .iter()
         .map(|&cardinality| StatRecord {
             cardinality,
-            cardinality_estimator: measure_error::<CardinalityEstimator<usize>>(cardinality),
+            cardinality_estimator: measure_error::<CardinalityEstimatorMut>(cardinality),
             amadeus_streaming: measure_error::<AmadeusStreamingEstimator>(cardinality),
             probabilistic_collections: measure_error::<ProbabilisticCollections>(cardinality),
             hyperloglog: measure_error::<HyperLogLog>(cardinality),
@@ -99,6 +99,15 @@ fn benchmark(c: &mut Criterion) {
         Table::new(results).with(table_config).to_string(),
     )
     .unwrap();
+}
+
+/// Cardinality estimator trait representing common estimator operations.
+trait CardinalityEstimatorTrait<T: Hash + ?Sized> {
+    fn new() -> Self;
+    fn insert(&mut self, item: &T);
+    fn estimate(&mut self) -> usize;
+    fn merge(&mut self, rhs: &Self);
+    fn name() -> String;
 }
 
 fn bench_insert<E: CardinalityEstimatorTrait<usize>>(
@@ -184,6 +193,30 @@ struct StatRecord {
     probabilistic_collections: String,
     hyperloglog: String,
     hyperloglogplus: String,
+}
+
+struct CardinalityEstimatorMut(CardinalityEstimator<usize>);
+
+impl CardinalityEstimatorTrait<usize> for CardinalityEstimatorMut {
+    fn new() -> Self {
+        Self(CardinalityEstimator::new())
+    }
+
+    fn insert(&mut self, item: &usize) {
+        self.0.insert(item);
+    }
+
+    fn estimate(&mut self) -> usize {
+        self.0.estimate()
+    }
+
+    fn merge(&mut self, rhs: &Self) {
+        self.0.merge(&rhs.0);
+    }
+
+    fn name() -> String {
+        "cardinality-estimator".to_string()
+    }
 }
 
 struct AmadeusStreamingEstimator(amadeus_streaming::HyperLogLog<usize>);
