+++ benches/requirements.txt
@@ -1,5 +1,4 @@
 chdb==1.1.0
-jupyterlab==4.0.9
 matplotlib==3.8.2
 numpy==1.26.2
 pandas==2.1.4
