+++ fuzz/fuzz_targets/serde.rs
@@ -1,6 +1,6 @@
 #![no_main]
 
-use cardinality_estimator::estimator::{CardinalityEstimator, CardinalityEstimatorTrait};
+use cardinality_estimator::estimator::CardinalityEstimator;
 use libfuzzer_sys::fuzz_target;
 
 fuzz_target!(|data: &[u8]| {
